apps/backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
aiofiles==24.1.0
    # via quart
aiohappyeyeballs==2.4.4
    # via aiohttp
aiohttp==3.11.11
    # via
    #   -r requirements.in
    #   microsoft-kiota-authentication-azure
aiosignal==1.3.2
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.7.0
    # via
    #   httpx
    #   openai
attrs==24.3.0
    # via aiohttp
azure-common==1.1.28
    # via azure-search-documents
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   microsoft-kiota-authentication-azure
azure-identity==1.19.0
    # via
    #   -r requirements.in
    #   msgraph-sdk
azure-search-documents==11.5.2
    # via -r requirements.in
azure-storage-blob==12.24.0
    # via -r requirements.in
beautifulsoup4==4.12.3
    # via
    #   markdownify
    #   markitdown
blinker==1.9.0
    # via
    #   flask
    #   quart
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.1
    # via
    #   markitdown
    #   pdfminer-six
    #   requests
click==8.1.8
    # via
    #   flask
    #   quart
cobble==0.1.4
    # via mammoth
cryptography==44.0.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pdfminer-six
    #   pyjwt
defusedxml==0.7.1
    # via youtube-transcript-api
deprecated==1.2.15
    # via
    #   opentelemetry-api
    #   opentelemetry-semantic-conventions
distro==1.9.0
    # via openai
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.0
    # via quart
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
gunicorn==23.0.0
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via
    #   httpx
    #   hypercorn
hpack==4.0.0
    # via h2
httpcore==1.0.7
    # via httpx
httpx[http2]==0.28.1
    # via
    #   microsoft-kiota-http
    #   msgraph-core
    #   openai
hypercorn==0.17.3
    # via quart
hyperframe==6.0.1
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.5.0
    # via opentelemetry-api
isodate==0.7.2
    # via
    #   azure-search-documents
    #   azure-storage-blob
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.5
    # via
    #   flask
    #   quart
jiter==0.9.0
    # via openai
joblib==1.4.2
    # via scikit-learn
lxml==5.3.0
    # via python-pptx
mammoth==1.8.0
    # via markitdown
markdownify==0.14.1
    # via markitdown
markitdown==0.0.1a3
    # via -r requirements.in
markupsafe==3.0.2
    # via
    #   jinja2
    #   quart
    #   werkzeug
microsoft-kiota-abstractions==1.6.7
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-authentication-azure==1.6.7
    # via
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-http==1.6.7
    # via
    #   msgraph-core
    #   msgraph-sdk
microsoft-kiota-serialization-form==1.6.7
    # via msgraph-sdk
microsoft-kiota-serialization-json==1.6.7
    # via msgraph-sdk
microsoft-kiota-serialization-multipart==1.6.7
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.6.7
    # via msgraph-sdk
msal==1.31.1
    # via
    #   -r requirements.in
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via
    #   -r requirements.in
    #   azure-identity
msgraph-core==1.1.8
    # via msgraph-sdk
msgraph-sdk==1.15.0
    # via -r requirements.in
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
numpy==2.2.1
    # via
    #   markitdown
    #   pandas
    #   scikit-learn
    #   scipy
openai==1.65.5
    # via
    #   -r requirements.in
    #   markitdown
openpyxl==3.1.5
    # via markitdown
opentelemetry-api==1.29.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-instrumentation==0.50b0
    # via -r requirements.in
opentelemetry-sdk==1.29.0
    # via
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.50b0
    # via
    #   opentelemetry-instrumentation
    #   opentelemetry-sdk
packaging==24.2
    # via
    #   gunicorn
    #   opentelemetry-instrumentation
pandas==2.2.3
    # via markitdown
pathvalidate==3.2.1
    # via markitdown
pdfminer-six==20240706
    # via markitdown
pendulum==3.0.0
    # via
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
pillow==11.0.0
    # via python-pptx
portalocker==2.10.1
    # via msal-extensions
priority==2.0.0
    # via hypercorn
propcache==0.2.1
    # via
    #   aiohttp
    #   yarl
puremagic==1.28
    # via markitdown
pycparser==2.22
    # via cffi
pydantic==2.10.6
    # via openai
pydantic-core==2.27.2
    # via pydantic
pydub==0.25.1
    # via markitdown
pyjwt[crypto]==2.10.1
    # via
    #   msal
    #   pyjwt
python-dateutil==2.9.0.post0
    # via
    #   microsoft-kiota-serialization-text
    #   pandas
    #   pendulum
    #   time-machine
python-dotenv==1.0.1
    # via -r requirements.in
python-pptx==1.0.2
    # via markitdown
pytz==2024.2
    # via pandas
quart==0.20.0
    # via
    #   -r requirements.in
    #   quart-cors
quart-cors==0.8.0
    # via -r requirements.in
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   azure-core
    #   markitdown
    #   msal
    #   tiktoken
    #   youtube-transcript-api
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.2
    # via scikit-learn
six==1.17.0
    # via
    #   azure-core
    #   markdownify
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
soupsieve==2.6
    # via beautifulsoup4
speechrecognition==3.12.0
    # via markitdown
std-uritemplate==2.0.1
    # via microsoft-kiota-abstractions
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.8.0
    # via -r requirements.in
time-machine==2.16.0
    # via pendulum
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   anyio
    #   azure-core
    #   azure-identity
    #   azure-search-documents
    #   azure-storage-blob
    #   openai
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   python-pptx
    #   speechrecognition
tzdata==2024.2
    # via
    #   pandas
    #   pendulum
urllib3==2.3.0
    # via
    #   microsoft-kiota-http
    #   requests
werkzeug==3.1.3
    # via
    #   flask
    #   quart
wrapt==1.17.0
    # via
    #   deprecated
    #   opentelemetry-instrumentation
wsproto==1.2.0
    # via hypercorn
xlsxwriter==3.2.0
    # via python-pptx
yarl==1.18.3
    # via aiohttp
youtube-transcript-api==0.6.3
    # via markitdown
zipp==3.21.0
    # via importlib-metadata
